Handle unlisted/undisclosed recipients
This handles headers like

To: undisclosed-recipients:;

Instead of considering the To header as empty/missing in this case (which
happens because it does not contains any valid addresses), the milter will
pretend the header contains the following address: "undisclosed". This allows
differentiating emails with no To header or a completely invalid To header,
and emails with an undisclosed recipients header.